+++ spark-performance-insight-core/src/main/java/com/fluffyeti/spark/performance/insight/service/TaskService.java
@@ -8,11 +8,11 @@ import org.springframework.stereotype.Service;
 @Service
 public class TaskService extends ServiceImpl<TaskMapper, GoldTaskModel> {
 
-    public long getExecutorCountForStage(String appId, int stageId) {
+    public long getExecutorCountForStage(String appId, long stageId) {
         return baseMapper.getExecutorCountForStage(appId, stageId);
     }
 
-    public long getExecutorCountForJob(String appId, int jobId) {
+    public long getExecutorCountForJob(String appId, long jobId) {
         return baseMapper.getExecutorCountForJob(appId, jobId);
     }
 }
